% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_size_tau_test_crv.R
\name{calc_size_tau_test_crv}
\alias{calc_size_tau_test_crv}
\title{Calculate the conservative required sample size for one-sided test}
\usage{
calc_size_tau_test_crv(target_tau, G0, G1, p1, sig_level = 0.05, power = 0.9)
}
\arguments{
\item{target_tau}{The value of tau measure under alternative}

\item{G0}{Censoring rate for control arm}

\item{G1}{Censoring rate for treatment arm}

\item{p1}{Assignment probability to treatment arm}

\item{sig_level}{Level of significance}

\item{power}{Expected power to attain}
}
\value{
A single number indicating the required sample size
}
\description{
Calculate the conservative required sample size for one-sided test
}
\examples{
calc_size_tau_test_crv(0.5, 0.1, 0.1, 0.5)
}
