% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc.R
\docType{data}
\name{pbc}
\alias{pbc}
\title{Mayo Clinic Primary Biliary Cholangitis Data}
\format{
\subsection{\code{pbc}}{

A data frame with 258 rows and 3 columns:
\describe{
\item{surv.time}{the survival of each subject in the trial (days)}
\item{event}{censoring indicator (1: dead; 0: censored)}
\item{arm}{treatment arm (1: D-penicillamine; 0: placebo)}
}
}
}
\source{
\url{https://cran.r-project.org/package=survival}
}
\usage{
pbc
}
\description{
This dataset is obtained from 'pbc' in package 'survival' by excluding the non-randomized individuals.
For background and details of the original dataset, please refer to the document page of 'survival'.
}
\keyword{datasets}
