% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tau_process.R
\name{plot.tau_process}
\alias{plot.tau_process}
\title{Plot the Tau Process with/without cure fraction}
\usage{
\method{plot}{tau_process}(x, ...)
}
\arguments{
\item{x}{an object of class "tau_process", returned by \code{tau_proc} function}

\item{...}{additional arguments passed to underlying plot method}
}
\value{
a list with components \code{x} and \code{y}.
}
\description{
This function plot the estimated tau process with/withour cure fraction obtained from \code{tau_proc}.
It can be used to monitor the progression of treatment effect (for susceptible subgroups).
}
\examples{
fit <- tau_proc(data = pbc)
plot(fit)
}
