% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summaryTauFit.R
\name{print.summaryTauFit}
\alias{print.summaryTauFit}
\title{Print method for summaryTauFit objects}
\usage{
\method{print}{summaryTauFit}(x, ...)
}
\arguments{
\item{x}{an object of class "summaryTauFit"}

\item{...}{additional arguments}
}
\description{
Print method for summaryTauFit objects
}
