% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.fit.R
\name{tau.fit}
\alias{tau.fit}
\title{Two-Sample inference based on local Kendall's tau}
\usage{
tau.fit(data, t)
}
\arguments{
\item{data}{a data.frame consisting of \code{arm}, \code{surv.time}, \code{event}.}

\item{t}{the specified truncation time.}
}
\value{
an object of class "tauFit" with components
\tabular{ll}{
\code{N0} \tab number of individuals with arm=0 \cr
\tab \cr
\code{N1} \tab number of individuals with arm=1 \cr
\tab \cr
\code{t} \tab the specified truncation time \cr
\tab \cr
\code{tau} \tab the estiamted local Kendall's tau measure \cr
\tab \cr
\code{var.r} \tab the estimated variance under random grouping design \cr
\tab \cr
\code{var.f} \tab the estimated variance under fixed grouping design \cr
}
}
\description{
calculate the necessary elements to do inference based on tau
}
\details{
The estimating formula for tau and variance are proposed by Yi-Cheng Tai, Weijing Wang and Martin T. Wells.
}
