% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_est.R
\name{tau_est}
\alias{tau_est}
\title{Estimate of local Kendall's tau with right-censored data}
\usage{
tau_est(arm, surv.time, event, t)
}
\arguments{
\item{arm}{a non-empty numeric vector of group indicators, coded as 0 or 1.}

\item{surv.time}{a non-empty numeric vector of survival times (possibly censored).}

\item{event}{the censoring indicator coded as 0 if censored; as 1 if failed}

\item{t}{the specified time point}
}
\value{
the estimated value of local Kendall's tau at specified time.
}
\description{
Estimate of local Kendall's tau with right-censored data
}
