% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_fit.R
\name{tau_fit}
\alias{tau_fit}
\title{Two-Sample inference based on local Kendall's tau}
\usage{
tau_fit(data, t)
}
\arguments{
\item{data}{a data.frame consisting of \code{arm}, \code{surv.time}, \code{event}.}

\item{t}{the specified truncation time.}
}
\value{
a list with components
\tabular{ll}{
\code{N0} \tab number of individuals with arm=0 \cr
\tab \cr
\code{N1} \tab number of individuals with arm=1 \cr
\tab \cr
\code{t} \tab the specified truncation time \cr
\tab \cr
\code{var.r} \tab estimated variance under random grouping design \cr
\tab \cr
\code{var.f} \tab estimated variance under fixed grouping design \cr
\tab \cr
\code{obj} \tab a list of components to \cr
\tab \cr
\code{class}
}
}
\description{
calculate the necessary elements to do inference based on tau
}
\details{
The estimating formula for tau and variance are proposed by Yi-Cheng Tai, Weijing Wang and Martin T. Wells.
}
