% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tauFit.R
\name{plot.tauFit}
\alias{plot.tauFit}
\title{Plot method for \code{tauFit} objects}
\usage{
\method{plot}{tauFit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{tauFit}, returned by \code{tau.fit} function.}

\item{...}{additional arguments passed to underlying plot method.}
}
\value{
a list with components \code{x} and \code{y}.
}
\description{
Plot method for \code{tauFit} objects
}
\examples{
"later"
}
