% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summaryTauFit.R
\name{print.summaryTauFit}
\alias{print.summaryTauFit}
\title{Print Method for "summaryTauFit" Objects}
\usage{
\method{print}{summaryTauFit}(x, ...)
}
\arguments{
\item{x}{an object of class "summaryTauFit"}

\item{...}{additional arguments passed to underlying \code{printCoefmat} method.}
}
\value{
None
}
\description{
user-friendly format to present the inference results obtained from \code{summary.tauFit}.
}
\examples{
fit <- tau.fit(data = pbc)
summary(fit)
}
