% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tauFit.R
\name{summary.tauFit}
\alias{summary.tauFit}
\title{Summarize the inference of local Kendall's tau}
\usage{
\method{summary}{tauFit}(obj, conf.int = 0.95)
}
\arguments{
\item{obj}{an object of class "tauFit"}

\item{conf.int}{the significance level of the confidence interval}
}
\value{
an object of class "summaryTauFit"
}
\description{
Summarize the inference of local Kendall's tau
}
\examples{
"Later"
}
