% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tauFit.R
\name{summary.tauFit}
\alias{summary.tauFit}
\title{Summarize the Inference Result of Tau Process at Last Specified Time}
\usage{
\method{summary}{tauFit}(object, conf.int = 0.95, ...)
}
\arguments{
\item{object}{an object of class "tauFit"}

\item{conf.int}{the significance level of the confidence interval}

\item{...}{additional arguments passed to underlying summary method}
}
\value{
an object of class "summaryTauFit"
}
\description{
This function summarizes the inference results obtained from \code{tau.fit}.
The results under random grouping design (complete randomization design) and fixed grouping design (random allocation rule / urn model) would be almost the same with large sample size.
}
\examples{
fit <- tau.fit(data = pbc)
summary(fit)
}
