% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_fit.R
\name{tau_fit}
\alias{tau_fit}
\title{Two-Sample inference based on local Kendall's tau}
\usage{
tau_fit(data, t)
}
\arguments{
\item{data}{a data.frame consisting of \code{arm}, \code{surv.time}, \code{event}}

\item{t}{the specified truncation time}
}
\value{
a list with components
}
\description{
Two-Sample inference based on local Kendall's tau
}
